build custom context provider
