+++ 11-effects-reducers-context/login-app_effects/src/App.js
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from "react";
+import React, { useContext} from "react";
 
 import Login from "./components/Login/Login";
 import Home from "./components/Home/Home";
@@ -6,41 +6,16 @@ import MainHeader from "./components/MainHeader/MainHeader";
 import AuthContext from "./context/auth-context";
 
 function App() {
-  const [isLoggedIn, setIsLoggedIn] = useState(false);
-
-  const storedUserLoggedInformation = localStorage.getItem("isLoggedIn");
-
-  useEffect(() => {
-    if (storedUserLoggedInformation === "1") {
-      setIsLoggedIn(true);
-    }
-  }, []);
-
-  const loginHandler = (email, password) => {
-    // We should of course check email and password
-    // But it's just a dummy/ demo anyways
-    localStorage.setItem("isLoggedIn", "1");
-    setIsLoggedIn(true);
-  };
-
-  const logoutHandler = () => {
-    localStorage.removeItem("isLoggedIn");
-    setIsLoggedIn(false);
-  };
+  const ctx = useContext(AuthContext);
 
   return (
-    <AuthContext.Provider
-      value={{
-        isLoggedIn,
-        onLogout: logoutHandler
-      }}
-    >
+    <>
       <MainHeader />
       <main>
-        {!isLoggedIn && <Login onLogin={loginHandler} />}
-        {isLoggedIn && <Home onLogout={logoutHandler} />}
+        {!ctx.isLoggedIn && <Login />}
+        {ctx.isLoggedIn && <Home />}
       </main>
-    </AuthContext.Provider>
+    </>
   );
 }
