+++ 11-effects-reducers-context/login-app_effects/src/context/auth-context.js
@@ -1,8 +1,45 @@
-import React from "react";
+import React, { useEffect, useState } from "react";
 
 const AuthContext = React.createContext({
   isLoggedIn: false,
-  onLogout: () => {}
+  onLogout: () => {},
+  onLogin: (mail, password) => {}
 });
+
+export const AuthContextProvider = (props) => {
+  const [isLoggedIn, setIsLoggedIn] = useState(false);
+
+  const storedUserLoggedInformation = localStorage.getItem("isLoggedIn");
+
+  useEffect(() => {
+    if (storedUserLoggedInformation === "1") {
+      setIsLoggedIn(true);
+    }
+  }, []);
+
+  const loginHandler = (email, password) => {
+    // We should of course check email and password
+    // But it's just a dummy/ demo anyways
+    localStorage.setItem("isLoggedIn", "1");
+    setIsLoggedIn(true);
+  };
+
+  const logoutHandler = () => {
+    localStorage.removeItem("isLoggedIn");
+    setIsLoggedIn(false);
+  };
+
+  return (
+    <AuthContext.Provider
+      value={{
+        isLoggedIn,
+        onLogout: logoutHandler,
+        onLogin: loginHandler
+      }}
+    >
+      {props.children}
+    </AuthContext.Provider>
+  )
+}
 
 export default AuthContext;
